Make timeline offset scrolling more stable
Minor change to waveform renderer

+++ src/include/grid_constants.h
@@ -1,6 +1,6 @@
 #pragma once
-#define MIN_ZOOM (1.0f / (float)(1 << 22))
-#define DEFAULT_ZOOM (1.0f / (float)(1 << 2))
+#define MIN_ZOOM (1.0 / (double)(1 << 22))
+#define DEFAULT_ZOOM (1.0 / (double)(1 << 2))
 #define GRID_WIDEST 0
 #define GRID_WIDE 1
 #define GRID_MED 2
